Fix bug when new user quits app while making first project
If a new user opened the "create new project" dialog but then quit the
app, their onboarding status would be updated to
"creating-first-project" and persisted to localStorage, but their
modal status would not be persisted to disk. On re-open, the app
would appear blank, because the user wasn't "brand-new" (and thus
didn't get the guppy welcome screen), but didn't have a modal
open (and thus did not get the create project modal, even though
their onboardingStatus was "creating-first-project").

There are several ways we can sidestep this; this commit is just
one possibility: consolidate the "creating first app" source of
truth into the modal status, which just reopens the Guppy welcome
page if a user quits while making their first project.